fix: fix tests and reorganize based on review
