+++ packages/backend/src/managers/modelsManager.spec.ts
@@ -520,7 +520,9 @@ describe('downloadModel', () => {
 
     mocks.onEventDownloadMock.mockImplementation(listener => {
       listener({
+        id: 'id',
         status: 'completed',
+        duration: 1000,
       });
     });
